requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
pydantic
pydantic-settings
asyncpg
sqlalchemy[asyncio]
alembic
redis[asyncio]
anthropic
openai-whisper
mediapipe
opencv-python-headless
numpy<2.0
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
timm
onnxruntime
opensmile
librosa
soundfile
spacy
transformers
ffmpeg-python
Pillow
boto3
python-jose[cryptography]
passlib[bcrypt]
httpx
structlog
openai
